Apply fix from 51937 to new minibar cell
